started splitting app js into smaller modules
